==> code/R/analysis/01_scenario_generation_solving.R <==
library(raster)
library(tidyverse)
library(magrittr)
library(foreach)
library(doParallel)
library(prioritizr)
library(Rfast)
library(ggpubr)
library(fasterize)
# setwd("E:/Richard/global_risk/")
# setwd("D:/Work/Papers/2019_global_risk/global_risk/")
library(here)
#library(SparseData)
# memory.limit(300000)

## Define functions
source(here("code/R/functions/multi-objective-prioritization.R"))
source(here("code/R/functions/perm.R"))
source(here("code/R/functions/calculate-targets.R"))

# parallelization
# n_cores <- 2
# cl <- makeCluster(n_cores)
# registerDoParallel(cl)

dr <- 100
data_resolution <- paste0(dr, "km2")

# what climate layer should be used
# clim <- "climate_frank_ehe.tif"
# clim <- "climate_pat_vocc_bio1.tif"
clim <- "climate_comb.tif"

pu <- raster(here("data/intermediate/", data_resolution, "land.tif"))
wdpa <- raster(here("data/intermediate/", data_resolution, "wdpa_terrestrial.tif"))
locked_in <- ifelse(!is.na(wdpa[][!is.na(pu[])]), TRUE, FALSE)


if( !file.exists(paste0("data/intermediate/", data_resolution, "/rij.rds"))){
  
  amph <- stack(list.files(here("data/raw/IUCN_AOH/Amph/"), full.names = TRUE))
  rij_amph <- rij_matrix(pu, amph)
  saveRDS(rij_amph, here("data/intermediate/", data_resolution, "rij_amph.rds"))
  
  bird <- stack(list.files(here("data/raw/IUCN_AOH/Bird/"), full.names = TRUE))
  rij_bird <- rij_matrix(pu, bird)
  saveRDS(rij_bird, here("data/intermediate/", data_resolution, "rij_bird.rds"))
  
  mamm <- stack(list.files(here("data/raw/IUCN_AOH/Mamm/"), full.names = TRUE))
  rij_mamm <- rij_matrix(pu, mamm)
  saveRDS(rij_mamm, here("data/intermediate/", data_resolution, "rij_mamm.rds"))
  
  rept <- stack(list.files(here("data/raw/IUCN_AOH/Rept/"), full.names = TRUE))
  rij_rept <- rij_matrix(pu, rept)
  saveRDS(rij_rept, here("data/intermediate/", data_resolution, "rij_rept.rds"))
  
  rij <- rbind(rij_amph, rij_bird, rij_mamm, rij_rept)
  
  features <- stack(amph, bird, mamm, rept)
  
  
  spec <- data.frame(id = 1:nlayers(features),
                     name = names(features),
                     stringsAsFactors = FALSE)
  
  saveRDS(rij, here("data/intermediate/", data_resolution, "rij.rds"))
  saveRDS(spec, here("data/intermediate/", data_resolution, "spec.rds"))
  
} else {
  rij <- readRDS(here("data/intermediate/", data_resolution, "rij.rds"))
  spec <- readRDS(here("data/intermediate/", data_resolution, "spec.rds"))
}


###
# rij_amph <- readRDS(here("data/intermediate/", data_resolution, "rij_amph.rds"))
# rij_bird <- readRDS(here("data/intermediate/", data_resolution, "rij_bird.rds"))
# rij_mamm <- readRDS(here("data/intermediate/", data_resolution, "rij_mamm.rds"))
# rij_rept <- readRDS(here("data/intermediate/", data_resolution, "rij_rept.rds"))
# 
# rij <- rbind(rij_amph, rij_bird, rij_mamm, rij_rept)

#for testing
# rij <- rij_amph

wb_mean <- raster(here("data/intermediate/", data_resolution, "wb_mean.tif"))
lands <- raster(here("data/intermediate/", data_resolution, "kehoe_land_system.tif"))
# clim_grid_ann <- raster(here("data/intermediate/", data_resolution, "probability-annual-iucn.tif"))
climr <- raster(here("data/intermediate/", data_resolution, clim))
# climr <- round(climr, 3)
###
# only keep values that are present in all 3 threat layers
###

cdf <- as.data.frame(stack(pu, wb_mean, lands, climr))
cdf_red <- cdf[!is.na(cdf$land), ]
keep <- !is.na(rowSums(cdf_red))

rij_mat <- as(rij, "TsparseMatrix")
rij_mat_red <- rij_mat[, keep]
rij_mat_use <- as(rij_mat_red, "dgCMatrix")

# total across all planning units
# for biodiversity features, values are % of cell occupied
# aoh = area of planning unit times total representation / 100
# set biodiversity targets
features <- rowSums(rij_mat_use) %>% 
  enframe(value = "aoh") %>% 
  # set target based on aoh
  mutate(prop0 = calculate_targets(aoh))

wb_val <- wb_mean[][!is.na(pu[])]
wb_val_red <- wb_val[keep]
wb_val_red <- (wb_val_red + min(wb_val_red)) * -1

lands_val <- lands[][!is.na(pu[])]
lands_val_red <- lands_val[keep]
lands_val_red <- (lands_val_red - 100) * -1

clim_val <- climr[][!is.na(pu[])]
clim_val_red <- clim_val[keep]
clim_val_red <- ((clim_val_red - min(clim_val_red))) + 0.01

locked_in_red <- locked_in[keep]

cost <- rbind(matrix(wb_val_red, nrow = 1),
              matrix(lands_val_red, nrow = 1),
              matrix(clim_val_red, nrow = 1),
              matrix(rep(1, length(wb_val_red)), nrow = 1)
)

runs <- tibble(name = list(c("S", "L", "C", "A")),
               scen = list(4,
                           1,
                           2,
                           3,
                           c(1,2),
                           c(2,1),
                           c(1,3),
                           c(3,1),
                           c(2,3),
                           c(3,2),
                           c(1,2,3),
                           c(1,3,2),
                           c(2,1,3),
                           c(2,3,1),
                           c(3,1,2),
                           c(3,2,1)),
               flip_priority = FALSE,
               gap = rep(0.1, 16))

dir.create(file.path(here("data", "final", paste0(data_resolution, "_", clim))), 
           showWarnings = FALSE)
runs_dir <- here("data", "final", paste0(data_resolution, "_", clim))
# gap <- 0.1
# 
# flip_priority <- FALSE
runs <- foreach(run = 1:nrow(runs), .combine = bind_rows) %do% {
  
  r <- runs[run, ]
  r$name_out <- paste(r$name[[1]][r$scen[[1]]], collapse = "")
  
  # if(!any(r$scen[[1]] %in% 3)){
  #   return(NULL)
  # }
  
  # Start the loop from index >1
  # if(run < 8) return(NULL)
  # paste(r$name[[1]][r$scen[[1]]], collapse = "")
  
  sprintf("Run : %s", paste(r$name[[1]][r$scen[[1]]], collapse = "")) %>% 
    message()
  
  cost_temp <- cost[r$scen[[1]],]
  gap_temp <- rep(r$gap, length(r$scen[[1]]))
  
  #solve multi objective function
  s_gur <- multiobjective_prioritization(rij = rij_mat_use,
                                         obj = cost_temp,
                                         pu_locked_in = locked_in_red,
                                         relative_target = features$prop0,
                                         gap = gap_temp,
                                         flip_priority = r$flip_priority,
                                         threads = parallel::detectCores() - 6)
  
  # save solution
  str_glue_data(r, "rds_run-", sprintf("%03d", run), "_s-{name_out}",
                "_gap-{gap}.rds") %>%
    file.path(runs_dir, .) %>%
    saveRDS(s_gur, .)
  
  rs1 <- raster(pu)
  rs1_val <- rs1[][!is.na(pu[])]
  
  rs1_val_red <- s_gur$solution
  rs1_val[keep] <- rs1_val_red
  rs1[][!is.na(pu[])] <- rs1_val
  
  str_glue_data(r, "solution_run-", sprintf("%03d", run), "_s-{name_out}",
                "_gap-{gap}.tif") %>% 
    file.path(runs_dir, .) %>% 
    writeRaster(rs1, overwrite = TRUE, .)
  
  rm(cost_temp, gap_temp, s_gur, rs1)
  r
  
}


##########
## Post Processing
##########
substrRight <- function(x, n){
  substr(x, nchar(x)-n+1, nchar(x))
}

land <- raster(here("data/intermediate/", data_resolution, "land.tif"))
land_area <- sum(land[], na.rm = TRUE) * prod(res(land))/1000000 /1000000


fls <- list.files(here("data", "final", paste0(data_resolution, "_", clim)), 
                  pattern = "*.tif$", full.names = TRUE)[1:16]
nms <- gsub(".tif$", "", fls) %>% 
  gsub(here("data/final", paste0(data_resolution, "_", clim),"solution_run-"), "", .)

nms2 <- str_split(gsub("_gap-0.1", "", nms), "_s-", simplify = T)[,2]

#gap 0.05 and flip == FALSE only
r_stack <- stack(fls)
names(r_stack) <- nms2



r_df <- as.data.frame(r_stack)

prot <- sum(locked_in_red) * dr / 1000000
(selected <- colSums(r_df, na.rm = TRUE) * dr /1000000)
selected - prot

(perc_tot <- round(selected/land_area * 100, 2))

(perc_no_prot <- round((selected - prot)/land_area * 100, 2))

(perc_increse <- perc_tot - prot/land_area*100)

data.frame(t(data.frame(perc_tot = round(perc_tot, 2),
                        perc_increse = round(perc_increse, 2)))) %>%
  write.csv(here("data", "final", paste0(data_resolution, "_", clim), "Table1.csv"))

out_sum <- sum(r_stack[[-1]])
values(out_sum)[values(out_sum) == 0] <-  NA

sum_tab <- table(out_sum[])
sum_tab[length(sum_tab) + 1] <- sum_tab[length(sum_tab)] - sum(locked_in_red)
names(sum_tab)[length(sum_tab)] <- "15-prot"

round(sum_tab * 100 / (land_area * 1000000) * 100, 2)[-15]


#biodiversity hotspot overlap
dr <- 100
data_resolution <- paste0(dr, "km2")

trm <- raster(here("data/raw/IUCN_AOH/Amph/Abavorana luctuosa_aoh_10km.tif"))
base_raster <- raster(trm) 
res(base_raster) <- sqrt(dr) * 1000

hotsp <- st_read(here("data/raw/Biod/hotspots_2016_1.shp")) %>%
  filter(Type == "hotspot area")
hotsp %<>% st_transform(crs = crs(base_raster))
hotsp$one <- 1

hotsp_rast <- fasterize(hotsp, base_raster, field = "one")

sel_10 <- out_sum > 10
sel_10_no_prot <- sum(stack(sel_10, wdpa*10), na.rm = TRUE) ==1

tbl_hotsp <- table(sum(stack(sel_10_no_prot, hotsp_rast * 10), na.rm = TRUE)[])

# ss <- sum(r_stack)
# tt <-table(round(ss[],0))
# tt[10] <- tt[9] - sum(locked_in_red)
# names(tt)[10] <- "8-prot"
# tt
# 
# writeRaster(r_stack, here("data/final/", data_resolution,"solution.tif"), bylayer = TRUE, suffix = 'names')


######
# Summarise by land use type? (maybe using Scott's simplified categories)
# summarise by country and report averages?

gadm_df <- read_csv(here("data/intermediate/", data_resolution, "gadm_country_tbl.csv")) %>%
  select(NAME_IDX, GID_0, NAME_0)

gadm_country <- raster(here("data/intermediate/", data_resolution, "gadm_country.tif"))

count_stack <- stack(gadm_country, r_stack)
count_df <- as.data.frame(count_stack) %>% drop_na() 

count_df %<>% left_join(gadm_df, by = c("gadm_country" = "NAME_IDX"))

count_sum <- count_df %>% group_by(NAME_0)  %>% summarise_at(2:(ncol(count_df)-2), list(sum = sum))

count_sum_t <- data.frame(t(as.data.frame(count_sum   )), stringsAsFactors = FALSE)
colnames(count_sum_t) <- as.character(unlist(count_sum_t[1,]))
count_sum_t <- count_sum_t[-1, ]
rwnms <- row.names(count_sum_t) %>%
  gsub("_sum", "", .)
count_sum_t <- mutate_all(count_sum_t, function(x) as.numeric(as.character(x)))
row.names(count_sum_t) <- rwnms
# %>% tally()

count_sum_t %>% write.csv(here("data", "final", paste0(data_resolution, "_", clim), "country_summaries.csv"))

n_cell_count <- count_df %>% group_by(NAME_0) %>% summarise(n=n())

round(t(t(count_sum_t) / n_cell_count$n * 100), 2) %>%
  write.csv(here("data", "final", paste0(data_resolution, "_", clim), "country_summaries_perc.csv"))


# clean up
# stopCluster(cl)

count_sum_2 <- data.frame(mapply('/', count_sum_t, count_sum_t[1,])[2:nrow(count_sum_t),])
row.names(count_sum_2) <- rwnms[2:nrow(count_sum_t)]

is.na(count_sum_2) <- sapply(count_sum_2, is.infinite)
count_sum_2[is.na(count_sum_2)] <- NA

count_sum_2 %>% 
  write.csv(here("data", "final", paste0(data_resolution, "_", clim), "Table2_detail.csv"))

c_sum_tab <- data.frame(scenario = row.names(count_sum_2),
                        mean = apply(count_sum_2, 1, function(x) mean(x, na.rm = TRUE)),
                        min = apply(count_sum_2, 1, function(x) min(x, na.rm = TRUE)),
                        max = apply(count_sum_2, 1, function(x) max(x, na.rm = TRUE)),
                        low = apply(count_sum_2, 1, function(x) quantile(x, prob = 0.05, na.rm = TRUE)),
                        high = apply(count_sum_2, 1, function(x) quantile(x, prob = 0.95, na.rm = TRUE)))

c_sum_tab %>% 
  write.csv(here("data", "final", paste0(data_resolution, "_", clim), "Table2.csv"), row.names = FALSE)


rng <- colMaxs(as.matrix(count_sum_2), value = T) - colMins(as.matrix(count_sum_2), value = T)
rng <- tibble(NAME_0 = colnames(count_sum_2), rng = rng)

count_df2 <- count_df %>% 
  mutate(name = NAME_0, 
         NAME_0 = make.names(NAME_0),
         governance = cost[1,] / max(cost[1,]),
         landsys = cost[2,] / max(cost[2,]),
         climate = cost[3,] / max(cost[3,])) %>%
  left_join(rng, by = "NAME_0")

count_df3 <- subset(count_df2, 
                    subset = !duplicated(count_df2[c("gadm_country")]),
                    select = c("gadm_country", "name", "rng")) %>%
  arrange(name)


gadm_country2 <- gadm_country 
gadm_df <- data.frame(gadm_country = values(gadm_country2)) %>%
  left_join(count_df3, by = "gadm_country")
gadm_country2[] <- gadm_df$rng


gg_df <- count_df2 %>%
  group_by(name) %>%
  summarise(cells = sum(!is.na(A)),
            governance = mean(governance, na.rm = T),
            landsys = mean(landsys, na.rm = T),
            climate = mean(climate, na.rm = T)) %>%
  inner_join( 
    count_df2 %>%
      group_by(name) %>%
      summarise_at(vars(c(A:CLS)), sum),
    by='name' ) %>%
  mutate(SA = S/A,
         LA = L/A,
         CA= C/A) %>%
  filter(is.finite(SA))
#remove NAN and inf

gs <- ggplot(gg_df, aes(x = governance, y = SA)) +
  geom_point(size = 1) +
  theme_classic() +
  xlab("Governance") + 
  ylab("Result variability") + 
  geom_smooth(method = "lm")


gl <- ggplot(gg_df, aes(x = landsys, y = LA)) +
  geom_point(size = 1) +
  theme_classic() +
  xlab("Land systems") + 
  ylab("") + 
  geom_smooth(method = "lm")

gc <- ggplot(gg_df, aes(x = climate, y = CA)) +
  geom_point(size = 1) +
  theme_classic() +
  xlab("Climate") + 
  ylab("") + 
  geom_smooth(method = "lm")

ggarrange(gs, gl, gc, ncol = 3, nrow = 1) %>%
  ggexport(filename = here("manuscript", paste0("figures", "_", clim), paste0("Figure 4", ".png")),
           width = 2500 , height = 1000 , res = 400)

is.na(gg_df) <- sapply(gg_df, is.infinite)
gg_df[is.na(gg_df)] <- NA


#############################################################################
#Fig3 without PA's
wdpa[is.na(wdpa)] <- 0

count_stack <- stack(gadm_country, r_stack, wdpa)
count_df <- as.data.frame(count_stack) %>% drop_na() 

count_df %<>% left_join(gadm_df, by = c("gadm_country" = "NAME_IDX"))

count_sum <- count_df %>% group_by(NAME_0)  %>% summarise_at(2:(ncol(count_df)-2), list(sum = sum))

count_sum_t <- data.frame(t(as.data.frame(count_sum   )), stringsAsFactors = FALSE)
colnames(count_sum_t) <- as.character(unlist(count_sum_t[1,]))
count_sum_t <- count_sum_t[-1, ]
rwnms <- row.names(count_sum_t) %>%
  gsub("_sum", "", .)
count_sum_t <- mutate_all(count_sum_t, function(x) as.numeric(as.character(x)))
row.names(count_sum_t) <- rwnms
# %>% tally()

for(ii in 1:16){
  count_sum_t[ii,] <-  count_sum_t[ii,] - count_sum_t[17,]
}


count_sum_2 <- data.frame(mapply('/', count_sum_t, count_sum_t[1,])[2:nrow(count_sum_t) -1 ,])
row.names(count_sum_2) <- rwnms[2:nrow(count_sum_t)]

is.na(count_sum_2) <- sapply(count_sum_2, is.infinite)
count_sum_2[is.na(count_sum_2)] <- NA

rng <- colMaxs(as.matrix(count_sum_2), value = T) - colMins(as.matrix(count_sum_2), value = T)
rng <- tibble(NAME_0 = colnames(count_sum_2), rng = rng)

count_df2 <- count_df %>% 
  mutate(name = NAME_0, 
         NAME_0 = make.names(NAME_0),
         governance = cost[1,] / max(cost[1,]),
         landsys = cost[2,] / max(cost[2,]),
         climate = cost[3,] / max(cost[3,])) %>%
  left_join(rng, by = "NAME_0")

count_df3 <- subset(count_df2, 
                    subset = !duplicated(count_df2[c("gadm_country")]),
                    select = c("gadm_country", "name", "rng")) %>%
  arrange(name)


gadm_country2 <- gadm_country 
gadm_df <- data.frame(gadm_country = values(gadm_country2)) %>%
  left_join(count_df3, by = "gadm_country")
gadm_country2[] <- gadm_df$rng
